+++ src/main/java/data/duke.txt
@@ -1,2 +1,2 @@
-[T][X] homework
 [T][ ] homework
+[D][X] do this (by: 2012-12-20)
