+++ src/main/java/duke/Duke.java
@@ -1,9 +1,15 @@
+package dukey;
 import java.util.ArrayList;
 import java.util.Scanner;
+import java.time.LocalDate;
+import dukey.TaskList;
+import dukey.Storage;
+import dukey.Todo;
+import dukey.Deadline;
+import dukey.Event;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Paths;
-import java.time.LocalDate;
 
 class Duke {
     public static void main(String[] args) {
@@ -159,192 +165,3 @@ enum CommandType {
     LIST, MARK, UNMARK, DELETE, TODO, DEADLINE, EVENT, UNKNOWN
 }
 
-class Task {
-    protected String description;
-    protected boolean isDone;
-
-    public Task(String description) {
-        this.description = description;
-        this.isDone = false;
-    }
-
-    public void markAsDone() {
-        isDone = true;
-    }
-
-    public void markAsNotDone() {
-        isDone = false;
-    }
-
-    @Override
-    public String toString() {
-        return "[" + (isDone ? "X" : " ") + "] " + description;
-    }
-}
-
-class Todo extends Task {
-    public Todo(String description) {
-        super(description);
-    }
-
-    @Override
-    public String toString() {
-        return "[T]" + super.toString();
-    }
-}
-
-class Deadline extends Task {
-    public LocalDate by;
-
-    public Deadline(String description, String by) {
-        super(description);
-        LocalDate date = LocalDate.parse(by);
-        this.by = date;
-    }
-
-    @Override
-    public String toString() {
-        return "[D]" + super.toString() + " (by: " + by + ")";
-    }
-}
-
-class Event extends Task {
-    public String from;
-    public String to;
-
-    public Event(String description, String from, String to) {
-        super(description);
-        this.from = from;
-        this.to = to;
-    }
-
-    @Override
-    public String toString() {
-        return "[E]" + super.toString() + " (from: " + from + " to: " + to + ")";
-    }
-}
-
-
-class TaskList {
-    private ArrayList<Task> tasks;
-
-    public TaskList(String input) {
-        this.tasks = new ArrayList<>();
-
-        String[] lines = input.split("\n");
-
-        for (String line : lines) {
-            if (line.equals("")) {
-                continue;
-            }
-            Task task = createTaskFromInput(line);
-            if (task != null) {
-                tasks.add(task);
-            }
-        }
-    }
-
-    public void add(Task task) {
-        tasks.add(task);
-    }
-
-    public void remove(int index) {
-        if (index >= 0 && index < tasks.size()) {
-            tasks.remove(index);
-        } else {
-            System.out.println("Invalid index.");
-        }
-    }
-
-    public Task get(int index) {
-        if (index >= 0 && index < tasks.size()) {
-            return tasks.get(index);
-        } else {
-            System.out.println("Invalid index.");
-            return null;
-        }
-    }
-
-    public int size() {
-        return tasks.size();
-    }
-    private Task createTaskFromInput(String input) {
-        char type = input.charAt(1);
-        boolean isDone = (input.charAt(4) == 'X');
-
-        int openParenIndex = input.indexOf('('); // Find the index of the open parenthesis
-
-        String description;
-
-        if (openParenIndex != -1) {
-            // If there is an open parenthesis, extract description till the open parenthesis
-            description = input.substring(7, openParenIndex).trim();  // Trim whitespace
-        } else {
-            // If there's no parenthesis, use the entire string after the type marker
-            description = input.substring(7).trim();  // Trim whitespace
-        }
-
-        Task task;
-
-        if (type == 'T') {
-            task = new Todo(description);
-        } else if (type == 'D') {
-            int byIndex = input.indexOf("(by: ");
-            String by = input.substring(byIndex + 5, input.length() - 1).trim();  // Trim whitespace
-            task = new Deadline(description, by);
-        } else if (type == 'E') {
-            int fromIndex = input.indexOf("(from: ");
-            int toIndex = input.indexOf("to: ");
-            String from = input.substring(fromIndex + 7, toIndex).trim();  // Trim whitespace
-            String to = input.substring(toIndex + 4, input.length() - 1).trim();  // Trim whitespace
-            task = new Event(description, from, to);
-        } else {
-            return null;
-        }
-
-        if (isDone) {
-            task.markAsDone();
-        }
-
-        return task;
-    }
-
-    @Override
-    public String toString() {
-        StringBuilder stringBuilder = new StringBuilder();
-        for (Task task : tasks) {
-            stringBuilder.append(task.toString());
-            stringBuilder.append("\n");
-        }
-        return stringBuilder.toString();
-    }
-
-}
-
-class Storage {
-    private String directory;
-
-    Storage(String directory) {
-        this.directory = directory;
-    }
-
-    public String load() {
-        String fileContent = "";
-        try {
-            byte[] encodedBytes = Files.readAllBytes(Paths.get(directory));
-            fileContent = new String(encodedBytes);
-        } catch (IOException e) {
-            System.out.println("Error reading file: " + e.getMessage());
-        }
-        return fileContent;
-    }
-
-    public void save(TaskList tasks) {
-        try {
-            String content = tasks.toString(); // Get the string representation of tasks
-            Files.write(Paths.get(directory), content.getBytes());
-        } catch (IOException e) {
-            System.out.println("Error writing to file: " + e.getMessage());
-        }
-    }
-}
