+++ src/main/java/duke/Moretask.java
@@ -1,0 +1,46 @@
+package dukey;
+import java.time.LocalDate;
+import dukey.Task;
+
+class Todo extends Task {
+    public Todo(String description) {
+        super(description);
+    }
+
+    @Override
+    public String toString() {
+        return "[T]" + super.toString();
+    }
+}
+
+class Deadline extends Task {
+    public LocalDate by;
+
+    public Deadline(String description, String by) {
+        super(description);
+        LocalDate date = LocalDate.parse(by);
+        this.by = date;
+    }
+
+    @Override
+    public String toString() {
+        return "[D]" + super.toString() + " (by: " + by + ")";
+    }
+}
+
+class Event extends Task {
+    public String from;
+    public String to;
+
+    public Event(String description, String from, String to) {
+        super(description);
+        this.from = from;
+        this.to = to;
+    }
+
+    @Override
+    public String toString() {
+        return "[E]" + super.toString() + " (from: " + from + " to: " + to + ")";
+    }
+}
+
