+++ src/main/java/duke/package.java
@@ -1,0 +1,7 @@
+package duke;
+
+class Duke {
+    public static void main(String[] args) {
+
+    }
+}
